WIP - introduce discount strategy introduce ThreeForTwoOffer
