+++ exercise/kotlin/day13/src/main/kotlin/santamarket/model/ChristmasElf.kt
@@ -1,20 +1,16 @@
 package santamarket.model
 
-import santamarket.model.offer.FiveForAmountOffer
-import santamarket.model.offer.Offer
-import santamarket.model.offer.TenPercentOffer
-import santamarket.model.offer.TwoForAmountOffer
+import santamarket.model.offer.*
 
 class ChristmasElf(private val catalog: SantamarketCatalog) {
     private val offers = mutableMapOf<Product, Offer>()
 
     fun addSpecialOffer(offerType: SpecialOfferType, product: Product, argument: Double) {
-        val offer =
-            when (offerType) {
+        val offer = when (offerType) {
                 SpecialOfferType.TEN_PERCENT_DISCOUNT -> TenPercentOffer(offerType, product, argument)
                 SpecialOfferType.TWO_FOR_AMOUNT -> TwoForAmountOffer(offerType, product, argument)
                 SpecialOfferType.FIVE_FOR_AMOUNT -> FiveForAmountOffer(offerType, product, argument)
-                else -> Offer(offerType, product, argument)
+                SpecialOfferType.THREE_FOR_TWO -> ThreeForTwoOffer(offerType, product, argument)
             }
         offers[product] = offer
     }
